refactor(ui): rename bottom-tab-navigator to bottom-navigation
